Improve account abstraction
Previously the baseaccount lived in x/auth while the coin lived in
types. The argument for coins to be in types is that it is fundamental
to almost every application. baseaccount is also fundamental to almost
every application and hence it should live in types.
